+++ nexus/lib/build.sh
@@ -21,8 +21,8 @@ cd -
 # build shared-library
 cp core/*.go tmp/
 cd tmp/
-go build -buildmode c-shared -o ../export/libwandbcore.so wandbcore.go nexusimage.go
-go build -buildmode c-archive -o ../export/libwandbcore.a wandbcore.go nexusimage.go
+go build -buildmode c-shared -o ../export/libwandbcore.so *.go
+go build -buildmode c-archive -o ../export/libwandbcore.a *.go
 mv ../export/libwandbcore.so ../export/libwandbcore.a ../export/lib/
 mv ../export/libwandbcore.h ../export/include/
 cd -
